+++ CHANGELOG.md
@@ -18,6 +18,11 @@
 - `singularity sign` now supports signing an image in an OCI-SIF with a
   cosign-compatible sigstore signature. Use the `--cosign` flag, and provide
   a private key with the `--key` flag.
+- `singularity verify` now supports verifying an image in an OCI-SIF with a
+  cosign-compatible sigstore signature. Use the `--cosign` flag, and provide a
+  public key with the `--key` flag. Verification passes if at least one
+  signature that can be validated with the provided key is present. The JSON
+  payloads of all valid signatures are displayed.
 
 ## Requirements / Packaging